Fixed docker-compose path, added separate package for types (not operational)
