+++ ydb/library/yql/providers/generic/connector/tests/datasource/clickhouse/ya.make
@@ -3,7 +3,8 @@ PY3TEST()
 STYLE_PYTHON()
 NO_CHECK_IMPORTS()
 
-DATA(arcadia/ydb/library/yql/providers/generic/connector/tests/datasource/clickhouse/docker-compose.yaml)
+DATA(arcadia/ydb/library/yql/providers/generic/connector/tests/datasource/clickhouse/docker-compose.yml)
+DATA(arcadia/ydb/library/yql/providers/generic/connector/tests/fq-connector-go)
 
 IF (AUTOCHECK) 
     # Split tests to chunks only when they're running on different machines with distbuild,
