+++ ydb/library/yql/providers/generic/connector/tests/test_cases/select_positive_clickhouse.py
@@ -6,7 +6,7 @@ from typing import Sequence
 from ydb.library.yql.providers.generic.connector.api.common.data_source_pb2 import EDataSourceKind, EProtocol
 from ydb.public.api.protos.ydb_value_pb2 import Type
 
-import ydb.library.yql.providers.generic.connector.tests.utils.clickhouse as clickhouse
+import ydb.library.yql.providers.generic.connector.tests.utils.types.clickhouse as clickhouse
 from ydb.library.yql.providers.generic.connector.tests.utils.schema import (
     Schema,
     Column,
