+++ ydb/library/yql/providers/generic/connector/tests/utils/ya.make
@@ -25,3 +25,7 @@ PEERDIR(
 )
 
 END()
+
+RECURSE_FOR_TESTS(
+    types
+)
