+++ ydb/library/yql/providers/generic/connector/tests/ya.make
@@ -1,0 +1,5 @@
+RECURSE_FOR_TESTS(
+    datasource
+    test_cases
+    utils
+)
